use bigint instead of numeric for keeping aggregate durations and do second -> hour conversion view

+++ datafiles/share/postgresql/sgas-postgres-3.4.2-3.5-upgrade.sql
@@ -140,8 +140,8 @@ CREATE TABLE uraggregated_data (
     runtime_environments_id integer[],
     status_id               integer,
     n_jobs                  integer,
-    cputime                 numeric(14,2),
-    walltime                numeric(14,2),
+    cputime                 bigint,
+    walltime                bigint,
     generate_time           timestamp
 );
